+++ src/main/scala/Item.scala
@@ -8,4 +8,4 @@ sealed trait TypeCategory extends Category
 case object Food extends TypeCategory
 case object Drink extends TypeCategory
 
-case class Item(name: String, price: Double, categories: Seq[Category])
+case class Item(name: String, price: BigDecimal, categories: Seq[Category])
